Feature/5 - picking up items (#55)
* rounded out last of inventory pickup system and fiddle with unity setting and .gitignore for vscode

* Added collider to stop movement and extended trigger to extend reach

Adds the ability to pick up items in the environment.

Added some test candies and ensured nothing is picked up unless the girl collides w/ an item.
Renamed "jump" to "action" in player settings.
Upgraded unity version and added .vscode to .gitignore
Added collider to the player
Moved adding logic out of the inventory system and into a new "PlayerAction" game object.
Updated the candy and player prefabs after all this shiznit
I made the Inventory.cs sort of a singleton (static method to grab the current instance) since there's only one inventory for the life of the game. I did this because the PlayerAction class needed to access the inventory. Let me know if this is bad form since its sort of a global variable.