Feature/57 - Inventory updates (#60)
* Feature/53 (#53)

* Player controller and clips added

* CharacterMovement script added

* Player prefab added and in scene.  No action functionality

* Feature/11 - Inventory System (#54)

* Added Canvas UI and Inventory Bar with one element

* Auto numbering of inventory slot logic

* Item Interface and example item created

* Can add items to inventory panel

* Can set currently selected item for use

* Can select inventory with numbers

* Feature/5 - picking up items (#55)

* rounded out last of inventory pickup system and fiddle with unity setting and .gitignore for vscode

* Added collider to stop movement and extended trigger to extend reach

Adds the ability to pick up items in the environment.

Added some test candies and ensured nothing is picked up unless the girl collides w/ an item.
Renamed "jump" to "action" in player settings.
Upgraded unity version and added .vscode to .gitignore
Added collider to the player
Moved adding logic out of the inventory system and into a new "PlayerAction" game object.
Updated the candy and player prefabs after all this shiznit
I made the Inventory.cs sort of a singleton (static method to grab the current instance) since there's only one inventory for the life of the game. I did this because the PlayerAction class needed to access the inventory. Let me know if this is bad form since its sort of a global variable.

* Fixed auto-navigation issue

* Refactored numbering

* Refactored back.  Numbering in InventorySlot

* Feature/12 - Flower Prefab (sprites, animator, and SFX) (#59)

* * Created flower sprite, animator, and SFX for the flower being created/growing.
* Added Audio folder.
* Added Animations folder.

* * Removed duplicated Script for FlowerGrowth
* Added flower to test scene

* Removed SPACE from Submit input in Input Manager

* Tested toggle instead of button

* Changed Slot into toggle instead of a button

* Player uses item when selected

* Removed (blocking) colliders from Candy

* Apertures added to use item in front of player with Space

* Updated Inventory in merged Scene